requirements.txt
asgiref==3.9.1
bleach==6.2.0
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
cryptography==46.0.2
cssselect2==0.8.0
Django==5.2.5
django-auditlog==3.2.1
django-cors-headers==4.7.0
django-filter==25.2
django-summernote==0.8.20.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-audit-trail==0.3.7
drf-extra-fields==3.7.0
drf-yasg==1.21.10
filetype==1.2.0
fonttools==4.60.0
google-auth==2.41.1
html5lib==1.1
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydyf==0.11.0
PyJWT==2.10.1
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3.post1
uvicorn
